// prototypes/antigravity-frontend/src/types/dataset.ts
export interface GeneratedImageAsset {
  path: string;
  width: number;
  height: number;
  sha256: string;
  sourceAssetPath: string;
}
// ...
export type PassiveUiStyleId =
  | "common"
  | "common_negative"
  | "rare_yellow"
  | "rare2_blue"
  | "world_tree";
// ...
export interface UiColorDefinition {
  linearR: number;
  linearG: number;
  linearB: number;
  alpha: number;
  cssHex: string;
}

export interface PassiveUiStyleDefinition {
  id: PassiveUiStyleId;
  nameKo: string;
  animationName: string;
  frameColor: UiColorDefinition;
  frameOverlayColor: UiColorDefinition;
  frameOverlayOpacity: number;
  backgroundColor: UiColorDefinition;
  textColor: UiColorDefinition;
  rankArrowColor: UiColorDefinition;
  gradientColor: UiColorDefinition;
  gradientOpacity: number;
  effectColor: UiColorDefinition;
  effectOpacity: number;
  prismOpacity: number;
  worldTreeEffectOpacity: number;
  evidenceKind: string;
}

export type WorldTreeUiLayerId =
  | "dissolve"
  | "scroll_slow"
  | "scroll_fast"
  | "base_gradient";

export interface WorldTreeUiLayerDefinition {
  id: WorldTreeUiLayerId;
  widgetName: string;
  materialAssetPath: string;
  color: UiColorDefinition;
  brushTintAlpha: number;
  textureKeys: string[];
}